Use array deduplicator for SearchProvider
